Функционал фильтров вынесен в отдельный модуль
