+++ js/modules/forms.js
@@ -26,7 +26,6 @@ const replaceResetHandler = (formSelector) => {
   const resetButton = form.querySelector('[type="reset"]');
   const filtersPanel = document.querySelector('.map__filters');
 
-
   resetButton.addEventListener('click', () => {
     const filtersResetButton = document.createElement('button');
     filtersResetButton.setAttribute('type', 'reset');
@@ -34,54 +33,8 @@ const replaceResetHandler = (formSelector) => {
     filtersPanel.appendChild(filtersResetButton);
     setTimeout(() => filtersResetButton.click(), 0);
     setTimeout(() => filtersResetButton.remove(), 0);
-
     setTimeout(resetMap, 0);
   });
 };
 
-const setFilterChangeHandler = (formSelector) => {
-  const form = document.querySelector(formSelector);
-  const filterInputs = form.querySelectorAll('input');
-  const filterSelects = form.querySelectorAll('select');
-
-  const getInputsValues = () => {
-    const featuresArray = [];
-    filterInputs.forEach((filter) => {
-      const value = filter.getAttribute('value');
-      if (filter.checked) {
-        featuresArray.push(value);
-      }
-    });
-    return featuresArray;
-  };
-
-  const getSelectsValues = () => {
-    const selectsValues = {};
-
-    filterSelects.forEach((select) => {
-      const name = select.getAttribute('name');
-      const value = select.value;
-      if (value !== 'any') {
-        selectsValues[name] = value;
-      };
-    });
-    return selectsValues;
-  };
-
-  filterInputs.forEach((input) => {
-    input.addEventListener('change', () => {
-      const featuresList = getInputsValues();
-      console.log(featuresList);
-    });
-    // console.log(filtersValues);
-  });
-
-  filterSelects.forEach((select) => {
-    select.addEventListener('change', () => {
-      const optionsList = getSelectsValues();
-      console.log(optionsList);
-    });
-  });
-};
-
-export { replaceSubmitHandler, replaceResetHandler, setFilterChangeHandler };
+export { replaceSubmitHandler, replaceResetHandler };
